fix: remove use of constants, improve error handling, null extensions
