+++ packages/graphql-language-service-server/src/GraphQLCache.ts
@@ -59,7 +59,7 @@ export async function getGraphQLCache(
 ): Promise<GraphQLCacheInterface> {
   let graphQLConfig = config || (await loadConfig({ rootDir: configDir }));
   if (extensions && extensions.length > 0) {
-    for (const extension of extensions) {
+    for await (const extension of extensions) {
       graphQLConfig = await extension(graphQLConfig);
     }
   }
@@ -633,7 +633,7 @@ export class GraphQLCache implements GraphQLCacheInterface {
       schema = await projectConfig.getSchema();
     }
 
-    const customDirectives = projectConfig.extensions.customDirectives;
+    const customDirectives = projectConfig?.extensions?.customDirectives;
     if (customDirectives && schema) {
       const directivesSDL = customDirectives.join('\n\n');
       schema = extendSchema(
