+++ packages/graphql-language-service-server/src/Logger.ts
@@ -69,8 +69,9 @@ export class Logger implements VSCodeLogger {
     const logMessage = `${timestamp} [${severity}] (pid: ${pid}) graphql-language-service-usage-logs: ${message}\n\n`;
     // write to the file in tmpdir
     fs.appendFile(this._logFilePath, logMessage, _error => {});
-    process.stderr.write(logMessage, err => {
-      console.error(err);
+    // const processSt = (severity === SEVERITY.ERROR) ? process.stderr : process.stdout
+    process.stderr.write(logMessage, _err => {
+      // console.error(err);
     });
   }
 }
