+++ packages/graphql-language-service-server/src/startServer.ts
@@ -66,72 +66,64 @@ type Options = {
  * @returns {Promise<void>}
  */
 export default async function startServer(options: Options): Promise<void> {
-  try {
-    const logger = new Logger();
+  const logger = new Logger();
 
-    if (options && options.method) {
-      let reader;
-      let writer;
-      switch (options.method) {
-        case 'socket':
-          // For socket connection, the message connection needs to be
-          // established before the server socket starts listening.
-          // Do that, and return at the end of this block.
-          if (!options.port) {
-            process.stderr.write(
-              '--port is required to establish socket connection.',
+  if (options && options.method) {
+    let reader;
+    let writer;
+    switch (options.method) {
+      case 'socket':
+        // For socket connection, the message connection needs to be
+        // established before the server socket starts listening.
+        // Do that, and return at the end of this block.
+        if (!options.port) {
+          process.stderr.write(
+            '--port is required to establish socket connection.',
+          );
+          process.exit(1);
+        }
+
+        const port = options.port;
+        const socket = net
+          .createServer(client => {
+            client.setEncoding('utf8');
+            reader = new SocketMessageReader(client);
+            writer = new SocketMessageWriter(client);
+            client.on('end', () => {
+              socket.close();
+              process.exit(0);
+            });
+            const connection = createMessageConnection(reader, writer, logger);
+            addHandlers(
+              connection,
+              logger,
+              options.configDir,
+              options?.extensions ?? [],
+              options.config,
             );
-            process.exit(1);
-          }
-
-          const port = options.port;
-          const socket = net
-            .createServer(client => {
-              client.setEncoding('utf8');
-              reader = new SocketMessageReader(client);
-              writer = new SocketMessageWriter(client);
-              client.on('end', () => {
-                socket.close();
-                process.exit(0);
-              });
-              const connection = createMessageConnection(
-                reader,
-                writer,
-                logger,
-              );
-              addHandlers(
-                connection,
-                logger,
-                options.configDir,
-                options.extensions,
-                options.config,
-              );
-              connection.listen();
-            })
-            .listen(port);
-          return;
-        case 'stream':
-          reader = new StreamMessageReader(process.stdin);
-          writer = new StreamMessageWriter(process.stdout);
-          break;
-        case 'node':
-        default:
-          reader = new IPCMessageReader(process);
-          writer = new IPCMessageWriter(process);
-          break;
-      }
-      const connection = createMessageConnection(reader, writer, logger);
-      addHandlers(
-        connection,
-        logger,
-        options.configDir,
-        options.extensions,
-        options.config,
-      );
-      connection.listen();
+            connection.listen();
+          })
+          .listen(port);
+        return;
+      case 'stream':
+        reader = new StreamMessageReader(process.stdin);
+        writer = new StreamMessageWriter(process.stdout);
+        break;
+      case 'node':
+      default:
+        reader = new IPCMessageReader(process);
+        writer = new IPCMessageWriter(process);
+        break;
     }
-  } catch (err) {
-    process.stderr.write(err.message);
+    const connection = createMessageConnection(reader, writer, logger);
+    addHandlers(
+      connection,
+      logger,
+      options.configDir,
+      options?.extensions ?? [],
+      options.config,
+    );
+    connection.listen();
   }
 }
